requirements.txt
google-genai>=2.0.0
python-dotenv>=1.0.0
edge-tts>=7.0.0
mutagen>=1.47.0
ffmpeg-python>=0.2.0
requests>=2.31.0
pyinstaller>=6.0.0
